Implemented all planned features, now moving on to code optimization and refactoring

+++ src/main/java/com/atri/service/impl/RecentRecordServiceImpl.java
@@ -3,7 +3,9 @@ package com.atri.service.impl;
 import com.atri.dao.Mapper;
 import com.atri.entity.Record;
 import com.atri.service.RecentRecordService;
+import com.atri.util.RecordDeletionException;
 import jakarta.annotation.Resource;
+import lombok.extern.java.Log;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -11,6 +13,7 @@ import java.time.LocalDateTime;
 import java.util.List;
 
 @Service
+@Log
 public class RecentRecordServiceImpl implements RecentRecordService {
 
     @Resource
@@ -22,22 +25,31 @@ public class RecentRecordServiceImpl implements RecentRecordService {
     }
 
     @Override
-    public int getMaxScore() {
+    public Integer getMaxScore() {
         return mapper.selectMaxScore();
     }
 
     @Override
     public int addRecentRecord(String role, int score) {
-        return mapper.insertRecentRecord(1, role, score, LocalDateTime.now());     // 将 user_id 默认设为 1
+        // 将 user_id 默认设为 1
+        return mapper.insertRecentRecord(1, role, score, LocalDateTime.now());
     }
 
-    @Transactional
+    @Transactional  // 使用实务操作，保证原子性
     @Override
     public int clearRecord() {
-        int deletedCount = mapper.deleteRecent();
-        mapper.resetAutoIncrement();
-        if (deletedCount == 0) {
-            throw new RuntimeException("删除记录失败qaq");
+        int deletedCount = 0;
+        try {
+            Integer maxScore = mapper.selectMaxScore();
+            if (maxScore == null) throw new RecordDeletionException("重复删除记录qaq");
+            deletedCount = mapper.deleteRecent();
+            mapper.resetAutoIncrement();
+        } catch (RecordDeletionException e) {
+            log.info(e.getMessage());
+        } catch (Exception e) {
+            log.warning(e.getMessage());
+            // 触发回滚
+            throw new RuntimeException("系统错误，删除失败qaq", e);
         }
         return deletedCount;
     }
